Added bintray provider, moved markdown conversion code to a separate file, removed useless line in the patreon provider.
